fix average calculation and change return type

+++ WebUI/Data/DataAccess/IReviewService.cs
@@ -15,7 +15,7 @@ namespace ConestogaInsidersClub.Data.DataAccess
         Task AddReview(Review review);
         Task DeleteReview(Review review);
         Task<List<Review>> GetReviewsByUser(int userId);
-        Task<int> GetAverageRating(int gameId);
+        Task<double> GetAverageRating(int gameId);
 
     }
 }
